consistent describe, merge column utils tests
